publish01

+++ Application/Shop/CartDto.cs
@@ -6,8 +6,8 @@ namespace Application.Shop
     {
         public long CartId { get; set; }
         public long Quantity { get; set; }
-        public double Price { get; set; }
-        public double Total { get; set; }
+        public decimal Price { get; set; }
+        public decimal Total { get; set; }
         public long Product { get; set; }
         public Guid Order { get; set; }
         public string ProductName { get; set; }
